Rename logging->logs to avoid name collision with std lib
